Keep track of photos that fail to load
I noticed that a good chunk of photos fail to load. This results in a
poor UX because new_round has to be called again which can take up an
additional 2-3 seconds.

To keep track of which photos are "bad", I created a Parse class called
ReportedMedia. It only has 2 columns:

  * mediaId (string): object ID of the photo
  * reports (number): number of times this photo has been reported

We should go through this list every 2-3 days and delete the bad photos.